Remove test-deployment logic from deployModule
This is a partial revert of 6a88fa7: Actual applications of the build
step have been shown to require very specific logic when (and under what
version-identifier) a testing (or stable) module has to be deployed.

This change therefore removes all logic involved in such decisions
from deployModule and makes the user responsible for handling the
deployment strategy for testing modules.

deployModule should be a generic step for simply deploying a module.
To make it useful for testing deployments, the default version string
can from now on be overwritten by a custom identifier (that may contain
changesets, patchsets, dates, commit hashes, ...)

Change-Id: I89987fb5c62ab0bd139e39577c49beb34ee7aef7